fix buggy interpreter

+++ uninstall.sh
@@ -1,2 +1,4 @@
-#!/sbin/sh
-rm -rf "/sdcard/Android/frpc" 2>/dev/null
+#!/system/bin/sh
+if [ -d "/sdcard/Android/frpc" ]; then
+	rm -rf "/sdcard/Android/frpc" 2>/dev/null
+fi
